Приложение протестировано на логику рабоыт

+++ store/models/completed_order.py
@@ -6,5 +6,5 @@ class CompletedOrders(db.Model):
     courier_id = db.Column(db.Integer, db.ForeignKey('courier.courier_id'))
     completed_orders = db.Column(db.Integer, default=0)
     last_complete_time = db.Column(db.DateTime)
-    general_complete_seconds = db.Column(db.Float)
+    general_complete_seconds = db.Column(db.DECIMAL)
     region = db.Column(db.Integer)
